add feature Service

+++ src/main/java/com/exult/ProjectCisco/repository/FeatureRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FeatureRepository extends JpaRepository<Feature,Integer> {
+public interface FeatureRepository extends JpaRepository<Feature,Long> {
     Feature findByName(String name);
 }
